fix: Move rodata section to the beginning
If gdt64_code_seg is defined after the place we refer it, there will be
an error said that "can't handle non absolute segment in `jmp'".

Because the assembler will see the ljmp and believe the symbol
gdt64_code_seg is an external symbol (that will be relocated by the
linker) and not a constant.

Notice that we just move the rodata section definition to the beginning
of the assembly source file, not to the beginning of the binary file,
which is controlled by linker script.

reference:

https://stackoverflow.com/questions/53934625/how-to-convert-the-following-nasm-code-into-att-syntax-so-as-to-fix-the-cannot